gemini-deep-research/pyproject.toml
[project]
name = "gemini-deep-research"
version = "0.1.0"
description = "Gemini Deep Research report generator"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.62.0",
]
